requirements.txt
tinytag
python-vlc
yt_dlp
validators
requests